requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
jupyter
ipykernel
matplotlib
seaborn
sodapy
ipdb
